reflexion point on how to reproduce dots
right now, the reproduce_dots_from_task_trial.m
is a good basis to extract data. The draw_dots()function is the crux.
It should be modified to only output the normalized dots positions.

Note how dealing with the change-point is a hard unsolved problem yet.

There is also the fira2dots.m file which was another, unfinished
attempt, to do the same thing as above.

I should try to merge the efforts from these two (three with
draw_dots) files.