fail horribly to convert .mat to .csv
current problem is that the location where I save the data locally
differs from the expected path from snow-dots.

An additional problem is that, somehow, the entries in the FIRA
data are not all of the same type (even though on the console
they look very much like scalars to me). So the automated FIRA
table production from snow-dots fails.

Not sure what to do. But prospects are grim.