try to fix previous bug
=== bug description ===

when calling load_single_file.m on session 2 from subject 1, an error was thrown at
the time of concatenating the FIRA table, and the error message said that the trialStart
variable was a cell in one table and a non-cell in the other table.
I pinpointed that this error occurred at the time of trying to concatenate data from Block2.

=== bug presumed cause ===

the trialStart and trialEnd fields of the trialData struct from the Quest block were empty vectors
([]) for many trials. Because of this, calling the function struct2table would result in a column
having the type of a cell.

=== bug fix ===

before calling struct2table, I replace []-values by nan.