each struct may be successfully converted to a table
in summary, the data from session 2019_06_20_13_45 has the following pbs:
1/ no group 'mainTopsTreeNode' exist in the topsDataLog object
2/ when collecting the data from all the groups with name 'trial_*' I notice that:
    2a) some skipped blocks appear as structs, and some others don't. Here, the block with taskID 3 is missing.
    2b) some blocks which appear as 'completed' in the subj_metadata.json file are completely missing. Here, the last block is completely missing.

I don't know if these issues are a peculiarity of this session only or not.
For the presumably missing block of data, I don't know if there is another way to access it, or if it really is missing! The latter would be worrying