finally things start to make sense
I finally managed to match entire blocks of trials from the data files
with the theoretical sequences planned.

The remaining issues that I am aware of are:

trialStart is NaN in the first row of many blocks, and sometimes in the
second row. I haven't looked at later rows when this happens.

trialEnd is NaN in a few datasets

dirChoice is NaN in a few datasets as well