count num NaN for trialStart, trialEnd, dirChoice
inconsistencies.txt is more readable now. It becomes more evident that
for some blocks, trialStart is NaN from the beginning, or sometimes
another crucial column. These blocks are probably the ones 'skipped'
during the experiment. For such blocks, it seems that the number of
trials is correctly reported in the meta_data file.