perf as fcn of coh for all subjects on Block2
Layout similar to metaplot.png

I am only plotting data for trials lasting 400 msec, and for blocks corresponding to the standard dots task instructions.
That is, the stimulus had no change-points, subjects knew it, and the task was to report the direction of motion (no change-point detection required).

------------------------Some info about the plots:-----------------------
one row per subject: the first three columns contain single-block data. The last column is data across all blocks (that meet the aforementioned criteria).
y-axis is always Prob(Correct); horizontal dashed lines signal 50% and 100% correct.

In each of the first three columns, each block was preceded by a Quest block that estimated the coherence threshold of the subject. The sequence of coherence values in the Quest block is represented in each inset. The golden psychometric function is the Weibull function, the slope of which was estimated in the Quest block. If the Quest estimation were perfect, the blue data points should lie on the golden line.

The little number floating to the right of each data point is the number of trials used to compute Prob(Correct). I computed the error bars in a Bayesian way, updating a Beta prior over P(Correct). I can give you more details about this next time we meet. Computing 99%-Confidence Intervals instead gives similar-looking intervals.

The title of each panel is the timestamp of the session

----------------------Potential issues that I notice are:---------------
The first subject, at 0%-coherence, consistently gets a high percent correct (total of 124 trials as shown in right-most panel). The only possible explanation that I can think of is that in our 0-coherence stimulus, there actually is some signal about the direction of motion of the dots, because my code had a 'coherenceSTD' parameter of 10... I will use these trials for the motion energy analysis, this low signal might help us bump the motion energy up, anyway, so it is not necessarily a bad thing.

For subject 1 again, conflating all the data into a single psychometric plot seems problematic (right-most panel), because P(Correct) is not a monotonically increasing function of coherence. Here, my intuition is that, since the third session for this subject was recorded on day 8, their perceptual threshold might have changed quite a bit.

Subject 5 had a horrible first Quest block. I know we said that we would re-run the block automatically in the future if such thing happens again. Nevertheless, I wonder whether that subject was simply not doing the task during the Quest block or really trying and failing...