donut plots for presenceCP vs. cpChoice
One thing that worries me is that the trial counts for probCP=0.2 at
coh=0 do not match with the Python plots :(

The other important thing is that, at 0 coherence, subjects are
clearly reluctant to report a CP

More improvement is needed on the aesthetics of these donut plots.
Maybe indicate number of trials in each strip, improve color,
improve legend, lay out plots in grid?