Compare frame counts task vs. dumped
the comparison is made in the frame_count.Rmd R notebook.
The goal is to check that the dots dumped as a matrix by
dDDKDebug have the same number of frames as the frame counters
stored in the task object (e.g. numberDrawPreCP in FIRA).

Answer is YES, frame counts match up.